refactor deposit fund endpoint and component
