+++ api/api/Controllers/CryptoController.cs
@@ -48,41 +48,6 @@ public class CryptoController : ControllerBase
         return Created("", result);
     }
 
-    [HttpPost("deposit-fund")]
-    public async Task<ActionResult<Response>> Depositfund([FromBody] DepositFundCommand command)
-    {
-        var userIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (string.IsNullOrEmpty(userIdClaim) || !int.TryParse(userIdClaim, out var userId))
-        {
-            return Unauthorized(new Response("Invalid user ID", false));
-        }
-
-        var commandWithUserId = command with { UserId = userId };
-        var result = await _mediator.Send(commandWithUserId);
-        if (!result.IsSuccess)
-        {
-            return BadRequest(result);
-        }
-        return Created(nameof(DepositFundCommand), result);
-    }
-
-    [HttpPost("withdraw-fund")]
-    public async Task<ActionResult<Response>> Withdrawfund([FromBody] WithdrawFundCommand command)
-    {
-        var userIdClaim = User.FindFirstValue(ClaimTypes.NameIdentifier);
-        if (string.IsNullOrEmpty(userIdClaim) || !int.TryParse(userIdClaim, out var userId))
-        {
-            return Unauthorized(new Response("Invalid user ID", false));
-        }
-
-        var commandWithUserId = command with { UserId = userId };
-        var result = await _mediator.Send(commandWithUserId);
-        if (!result.IsSuccess)
-        {
-            return BadRequest(result);
-        }
-        return Created(nameof(DepositFundCommand), result);
-    }
 
     [HttpGet("list-assets")]
     public async Task<ActionResult> ListCryptoAssets([FromQuery] ListCryptoAssetsQueryCommand command)
